requirements.txt
streamlit==1.31.0
pdfplumber==0.10.4
pymupdf>=1.24.0
pytesseract==0.3.10
Pillow>=10.3.0
opencv-python==4.9.0.80
sentence-transformers==2.3.1
chromadb>=0.4.22
google-generativeai==0.3.2
python-dotenv==1.0.1
rank-bm25==0.2.2
numpy>=1.26.0
pandas>=2.2.0

